+++ server/router/router.go
@@ -1,17 +1,11 @@
 package router
 
 import (
-	"bytes"
-	"io/ioutil"
 	"main/manager"
-	"main/pubsub"
-	"main/pubsub/systemevent"
 	"net/http"
-	"time"
 
 	"github.com/labstack/echo/v4"
 	"github.com/rakyll/statik/fs"
-	"github.com/ymotongpoo/goltsv"
 )
 
 type router struct {
@@ -56,18 +50,6 @@ func (r *router) getStaticHandler() func(c echo.Context) error {
 
 	h := http.FileServer(statikFs)
 	return echo.WrapHandler(http.StripPrefix("/", h))
-}
-
-func (r *router) getRawLogs(c echo.Context) error {
-	data, err := ioutil.ReadFile(`./accessLog`)
-	if err != nil {
-		pubsub.SystemEvent.Pub(pubsub.System{Time: time.Now(), Type: systemevent.ERROR, Message: err.Error()})
-	}
-	b := bytes.NewBufferString(string(data))
-
-	reader := goltsv.NewReader(b)
-	records, _ := reader.ReadAll()
-	return c.JSON(200, records) //内部的にJSON.marshalしているらしい
 }
 
 func (r *router) defaultBackend(c echo.Context) error {
